Fixed Product price data format> int to double

+++ data/src/main/java/com/productsapp/data/sources/remote/models/product/dummyJson/ProductDTO.kt
@@ -7,7 +7,7 @@ data class ProductDTO(
     val discountPercentage: Double = 0.0,
     val id: Int = 0,
     val images: List<String> = listOf(),
-    val price: Int = 0,
+    val price: Double = 0.0,
     val rating: Double = 0.0,
     val stock: Int = 0,
     val thumbnail: String = "",
